+++ src/rebar3_grisp.erl
@@ -18,5 +18,6 @@ init(State) ->
         rebar3_grisp_configure,
         rebar3_grisp_package,
         rebar3_grisp_version,
-        rebar3_grisp_report
+        rebar3_grisp_report,
+        rebar3_grisp_firmware
     ]).